Correction of warnings
Former-commit-id: 479a2147a54918e022eb669a9f47ed7b154a64f7

+++ src/server/game/SpellAuras.cpp
@@ -5845,7 +5845,7 @@ void Aura::CleanupTriggeredSpells()
     // needed for spell 43680, maybe others
     // TODO: is there a spell flag, which can solve this in a more sophisticated way?
     if(m_spellProto->Effects[GetEffIndex()].ApplyAuraName == SPELL_AURA_PERIODIC_TRIGGER_SPELL &&
-            m_spellProto->GetDuration() == m_spellProto->Effects[GetEffIndex()].Amplitude) // Non-trivial conversion
+            (uint32) m_spellProto->GetDuration() == m_spellProto->Effects[GetEffIndex()].Amplitude)
         return;
     m_target->RemoveAurasDueToSpell(tSpellId);
 }
